SF [ 1860228 ] LINUX: Even More GCC Warnings Fixes (thanks monkeyman)
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport@11127 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/lib/UnrarXLib/file.cpp
@@ -344,7 +344,7 @@ void File::Write(const void *Data,int Size)
 #if !defined(_WIN_32) && !defined(_LINUX)
         clearerr(hFile);
 #endif
-      if (Written<Size && Written>0)
+      if (Written<(unsigned int) Size && Written>0)
           Seek(Tell()-Written,SEEK_SET);
         continue;
       }
